requirements.txt
fastapi
uvicorn
python-dotenv
openai
langgraph
langchain-openai
requests

